Check if authenticated user can delete screenshot [ch340]

+++ app/Http/Livewire/ScreenshotHub/ShowShowPage.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire\ScreenshotHub;
 
 use App\Models\Screenshot;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class ShowShowPage extends Component
@@ -16,6 +17,10 @@ class ShowShowPage extends Component
 
     public function delete()
     {
+        if (Auth::id() !== $this->screenshot->user_id && Auth::user()->cannot('manage users')) {
+            abort(403, 'You don\'t have permission to delete this screenshot.');
+        }
+
         $this->screenshot->delete();
 
         session()->flash('alert', ['type' => 'success', 'message' => 'Screenshot successfully deleted!']);
